refactor(tile): move `TileKey` out of `cache.rs` into `tile/key.rs`
Every layer below cache (fetch backends, eventually the decoder lane)
speaks in `TileKey`s, so having the type live inside `cache.rs` made
`fetch/*.rs` reach back into a sibling module just for the address
type. With the upcoming three-layer pipe (backend → decoder → cache),
the cache module is one of several consumers — not the owner.

Move it to its own `tile/key.rs` and re-export from `tile/mod.rs`.
Importers updated:
- `fetch/mod.rs`, `fetch/http.rs` switch from `cache::TileKey` to
  `key::TileKey`.
- The `test_tile_key_display` test moves with the type and is renamed
  `display_uses_zxy_form` to match the new module's idiom.

Behavior change: zero. Pure type relocation. Second of three
mechanical prereqs (after the http rename, before the decode.rs
split) before the substantive `TileFetcher` / `FetchLane` redesign.
