+++ src/map/tile/mod.rs
@@ -1,6 +1,7 @@
 //! Tile subsystem — fetching, caching, decoding, and view calculations.
 //!
 //! Responsibilities:
+//!   key.rs       — `TileKey` (z, x, y) universal address
 //!   cache.rs     — memory + disk storage, decode, stale detection, prefetch
 //!   fetch/       — backends that populate the cache (HTTP today, more later)
 //!   decode.rs    — MVT protobuf decoding
@@ -9,8 +10,10 @@
 pub mod cache;
 pub mod decode;
 pub mod fetch;
+pub mod key;
 pub mod property;
 
 pub use cache::TileCache;
 pub use decode::Feature;
+pub use key::TileKey;
 pub use property::PropertyValue;
